Detect method definition from modules via include or extend
With the previous implementation method_added wouldn't catch methods that were
added to the class from a module via include or extend. In order to overwrite
the original method implementation as soon as it is defined the include and
extend methods are now defined by MetaMeta and check for methods as soon as they
are called.